+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Changelog may include unreleased versions.
 See [version roadmap](https://www.graalvm.org/release-notes/version-roadmap/) for release dates.
 
 ## Version 22.3.0
-* Implemented the [WebAssembly multi-value](https://github.com/WebAssembly/multi-value) proposal. It can be disabled using the `--js.wasm-multi-value=false` option;
+* Implemented the [WebAssembly multi-value](https://github.com/WebAssembly/multi-value) proposal.
 * Added an experimental option `--js.unhandled-rejections=handler` that allows to use a custom callback to track unhandled promise rejections.
 * Updated Node.js to version 16.16.0.
 * ECMA-402 Internationalization API has been enabled by default. It can be disabled by `--js.intl-402=false` option.